Update script.js
In the F9 Menu the thermal and the night vision were not working. It kept giving an error in the console ( CrBrowserMain/ Uncaught ReferenceError: actionName is not defined (@fireac/ui/js/script.js:375)" The code had actionname I fixed it with the correct code and now it works with no issues.